updater QoL modified:   update.py

+++ update.py
@@ -1,7 +1,7 @@
 import requests
 import os
 print("Welcome to the Update Wizard. This will update your current LB version.")
-name = input("What is the name of your LigmaBalls file? (DO NOT INCLUDE .PY)")
+name = input("What is the name of your LigmaBalls file? (DO NOT INCLUDE .PY) ")
 os.remove(name + ".py")
 url = 'https://github.com/Freakybob-Team/lb/blob/main/lb.py?raw=true'
 response = requests.get(url)
@@ -12,5 +12,9 @@ if response.status_code == 200:
         file.write(response.content)
         print('File downloaded successfully')
 print("Done! Your LigmaBalls version has been updated.")
-os.remove("update.py")
+wantremove = input("Do you want to delete update.py? (Yes/No)")
+if (wantremove == "Yes"):
+    os.remove("update.py")
+else:
+    print("Okay! We have not deleted update.py.")
 exec(open('lb.py').read())
